feat: v2.3.4 — per-action rate limits, brightness/volume debounce, callback race protection, readiness gating, sync diff
- Per-action cooldown overrides (cooldown_seconds_default + cooldown_overrides dict)
- Brightness/volume debounce coalesces rapid taps into single HA call (0.25s window)
- Per-user asyncio.Lock + idempotency guard for callback race protection
- _wait_for_ha requires both get_config() AND registry.sync() success
- Refresh button shows contextual diff (added/removed areas, devices, entities)
- 18 new tests, 95 total passing

+++ ha-telegram-bot/app/storage.py
@@ -200,7 +200,7 @@ class Database:
     # --- cooldown ---
 
     async def check_and_update_cooldown(
-        self, user_id: int, action: str, cooldown_seconds: int
+        self, user_id: int, action: str, cooldown_seconds: float
     ) -> tuple[bool, float]:
         assert self._db is not None
         now = time.time()
